block inapropriate access in the controllre and related tests

+++ app/controllers/articles_controller.rb
@@ -16,6 +16,10 @@ class ArticlesController < ApplicationController
   
   def edit 
     #@article = Article.find(params[:id])
+    unless @article.user == current_user
+    flash[:alert] = "You can only edit your own article."
+    redirect_to root_path
+    end
   end
   
   def create
@@ -34,21 +38,31 @@ class ArticlesController < ApplicationController
   
   def update
     #@article = Article.find(params[:id])
-    if @article.update(article_params)
-      flash[:success] = "Article has been updated"
-      redirect_to @article
+    unless @article.user == current_user
+      flash[:alert] = "You can only edit your own article."
+      redirect_to root_path
     else
-      flash.now[:danger] = "Article has not been updated"
-      render :edit
+      if @article.update(article_params)
+        flash[:success] = "Article has been updated"
+        redirect_to @article
+      else
+        flash.now[:danger] = "Article has not been updated"
+        render :edit
+      end
     end
   end
   
   def destroy
-    #@article = Article.find(params[:id]) 
+    #@article = Article.find(params[:id])
+    unless @article.user == current_user
+      flash[:alert] = "You can only delete your own article."
+      redirect_to root_path
+      else
       if @article.destroy
         flash[:success] = "Article has been deleted."
         redirect_to articles_path 
       end
+    end
   end
   
   protected
